v0.1.1 and v0.2.0
v0.1.1 was an attempt to, from scratch, use selectors of increasing specificity in a very organized manner to encompass every element on the page, while retaining the visually distinguishing features that I thought made the page look "cooler". It didn't get everything, though, as a lot of elements in other pages were not properly modified.

v0.2.0 was another from-scratch attempt, this time simply applying massive base changes (text to white, background to black) with few small, specific additional selectors for cleaning up anything that looked bad. The wide base modifiers captured everything that needed to be modified, and this approach seems to be working well due to the very very few number of elements that need to be directly modified after the major modifiers.

v0.2.0 is not as visually distinguished as v0.1.x but keeping it simple isn't necessarily a bad thing.